+++ sharding-jdbc-orchestration/src/main/java/io/shardingsphere/jdbc/orchestration/internal/state/datasource/DataSourceListenerManager.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.jdbc.orchestration.internal.state.datasource;
 
 import io.shardingsphere.core.event.ShardingEventBusInstance;
-import io.shardingsphere.core.orche.eventbus.state.disabled.DisabledStateEventBusEvent;
+import io.shardingsphere.core.orche.eventbus.state.DisabledStateEventBusEvent;
 import io.shardingsphere.jdbc.orchestration.internal.listener.ListenerManager;
 import io.shardingsphere.jdbc.orchestration.internal.state.StateNode;
 import io.shardingsphere.jdbc.orchestration.reg.api.RegistryCenter;
