+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/core/datasource/MasterSlaveDataSource.java
@@ -28,7 +28,7 @@ import io.shardingsphere.core.jdbc.core.connection.MasterSlaveConnection;
 import io.shardingsphere.core.orche.datasource.CircuitBreakerDataSource;
 import io.shardingsphere.core.orche.eventbus.config.MasterSlaveConfigurationEventBusEvent;
 import io.shardingsphere.core.orche.eventbus.state.CircuitStateEventBusEvent;
-import io.shardingsphere.core.orche.eventbus.state.disabled.DisabledStateEventBusEvent;
+import io.shardingsphere.core.orche.eventbus.state.DisabledStateEventBusEvent;
 import io.shardingsphere.core.rule.MasterSlaveRule;
 import lombok.Getter;
